Stocks_Info_Web_Application/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
requests==2.27.1
python-dateutil==2.8.2
